Add territory filters for solicitaton
