+++ app/controllers/reminders/reminders_controller.rb
@@ -1,5 +1,7 @@
 module Reminders
   class RemindersController < ApplicationController
+    include TerritoryFiltrable
+
     before_action :authenticate_admin!
     before_action :maybe_review_expert_subjects
 
@@ -19,23 +21,6 @@ module Reminders
           rejected: needs.diagnosis_completed.rejected.size
         }
       end
-    end
-
-    def retrieve_territory
-      safe_params = params.permit(:territory)
-      territory_id = safe_params[:territory] || session[:territory]
-      if territory_id.present?
-        session[:territory] = territory_id
-        Territory.find(territory_id)
-      else
-        session.delete(:territory)
-        nil
-      end
-    end
-
-    def find_territories
-      @territories = Territory.regions.order(:name)
-      @territory = retrieve_territory
     end
   end
 end
